refactor: Moves ppe-related routes into ppe.js
